Add Token routes and autologin and logout

+++ src/store/cart/action.js
@@ -1,18 +1,26 @@
+
+import setHeaders from '../reducer'
 export const FETCH_CART='FETCH_CART';
 export const ADD_TO_CART='ADD_TO_CART';
 export const DELETE_FROM_CART='DELETE_FROM_CART';
 
 
 const cart={items:[],totalAmount:0}
-const token='eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InZhbnNoZG9uOTkjZ21haWwuY29tIiwidXNlcklkIjoiNWYxYTA1YTBlMWI2MjgyZDJjZTIyNzBlIiwiaWF0IjoxNTk1NjMyMjc4LCJleHAiOjE1OTU2NDMwNzh9.C8WXGVXe4KbGqBOcK7y62QCUW7tvHLvqdJV0JjdRqls'
+let x = document.cookie;
+
+const token=`${x.split('=')[1]}`
+
 export const fetchCart=()=>{
+
+
+    if(!token){
+        return
+    }
+   
     return (dispatch)=>{
         fetch('http://localhost:3001/cart',{
             method:'get',
-            headers:{
-                'Authorization':token
-               
-            }
+            headers:setHeaders({'Content-Type': 'application/json'})
         }).then(response=>{
            return response.json()
         }).then(Cart=>{
@@ -28,6 +36,10 @@ export const fetchCart=()=>{
 }}
 export const AddToCart=(product,qty,Amt)=>{
 
+
+    if(!token){
+        return
+    }
     const Product={...product};
     console.log(Product)
     const Amts=+Amt 
@@ -37,10 +49,7 @@ export const AddToCart=(product,qty,Amt)=>{
         fetch('http://localhost:3001/cart/'+Product._id,{
             method:'post',
             body:JSON.stringify({Amt:Amts,qty,MRP:Product.MRP,sellingPrice:Product.sellingPrice}),
-            headers:{
-                'Authorization':token,
-                'Content-Type': 'application/json'
-            },
+            headers:setHeaders({'Content-Type': 'application/json'})
         }).then(response=>{
            return response.json()
         }).then(Cart=>{
@@ -53,15 +62,16 @@ export const AddToCart=(product,qty,Amt)=>{
     }
 }
 export const DeleteFromCart=(id)=>{
-    
+
+
+    if(!token){
+        return
+    }
     return (dispatch)=>{
         fetch('http://localhost:3001/cart/'+id,{
             method:'delete',
             body:JSON.stringify({id}),
-            headers:{
-                'Authorization':token,
-                'Content-Type': 'application/json'
-            },
+            headers:setHeaders({'Content-Type': 'application/json'}),
         }).then(response=>{
            return response.json()
         }).then(Cart=>{
