+++ src/store/order/action.js
@@ -1,18 +1,23 @@
+import setHeaders from '../reducer'
 export const FETCH_ORDER='FETCH_ORDER';
 export const ADD_ORDER='ADD_ORDER';
 export const REMOVE_ORDER='REMOVE_ORDER';
 export const FETCH_ORDER_DETAIL='FETCH_ORDER_DETAIL';
 
-const token='eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InZhbnNoZG9uOTkjZ21haWwuY29tIiwidXNlcklkIjoiNWYxYTA1YTBlMWI2MjgyZDJjZTIyNzBlIiwiaWF0IjoxNTk1NjMyMjc4LCJleHAiOjE1OTU2NDMwNzh9.C8WXGVXe4KbGqBOcK7y62QCUW7tvHLvqdJV0JjdRqls'
+let x = document.cookie;
+
+const token=`${x.split('=')[1]}`
+
 
 const Orders=[{id:0,title:'Banana',price:20,qty:5,date:'12-07-2019',status:'Confirm'},{id:1,title:'Chips',price:10,qty:3,date:'1-10-2019',status:'Cancel'},{id:2,title:'Chocolate',price:100,qty:1,date:'1-10-2019',status:'Delivered'}]
 export const fetchOrder=()=>{
+    if(!token){
+        return
+    }
     return dispatch=>{
         fetch('http://localhost:3001/order',{
-            headers:{
-                'Authorization':token,
-               
-            }
+            headers:setHeaders({'Content-Type': 'application/json'})
+
         })
         .then(res=>{
             return res.json()
@@ -25,6 +30,9 @@ export const fetchOrder=()=>{
     }
 }
 export const fetchDetail=(id)=>{
+    if(!token){
+        return
+    }
     const order=Orders.find((order)=>order.id===id)
     return dispatch=>{
             setTimeout(()=>{
@@ -33,15 +41,16 @@ export const fetchDetail=(id)=>{
     }
 }
 export const AddOrder=(orderData)=>{
+    if(!token){
+        return
+    }
     console.log(orderData)
     return dispatch=>{
             fetch("http://localhost:3001/order",{
                 method:'post',
                 body:JSON.stringify(orderData),
-                headers:{
-                    'Authorization':token,
-                    'Content-Type': 'application/json'
-                }
+                headers:setHeaders({'Content-Type': 'application/json'})
+
             }).then(res=>{
                 return res.json()
             }).then(order=>{
@@ -53,14 +62,14 @@ export const AddOrder=(orderData)=>{
     }
 }
 export const deleteOrder=(id)=>{
-    console.log(id)
+    if(!token){
+        return
+    }
     return dispatch=>{
             fetch("http://localhost:3001/order/"+id,{
                 method:'delete',
-                headers:{
-                    'Authorization':token,
-                    'Content-Type': 'application/json'
-                }
+                headers:setHeaders({'Content-Type': 'application/json'})
+
             }).then(res=>{
                 return res.json()
             }).then(order=>{
